optimizing

+++ src/app.ts
@@ -4,7 +4,9 @@ import fs from "fs";
 
 async function main() {
     const client = new Client();
-    fs.writeFileSync("log.txt", 'time;value\r\n');
+    if (!fs.existsSync("log.txt")) {
+        fs.writeFileSync("log.txt", 'time;value\r\n');
+    }
     client.getDevice({ host: '192.168.0.198' }).then((device: AnyDevice) => {
 
         setInterval(() => {
@@ -13,7 +15,7 @@ async function main() {
                 const powerValue = workaroundForStupids["emeter"]["realtime"]["power"];
                 console.log(`Writing ${powerValue} to file log.txt`);
                 fs.appendFileSync("log.txt", `${new Date().toISOString()};${powerValue}\r\n`);
-            });
+            }).catch(error => console.log(error));
         }, 1000);
     });
 }
